requirements.txt
streamlit==1.32.0
Pillow==10.2.0
docxtpl==0.16.7
python-docx==1.1.0
google-generativeai==0.4.1
python-dotenv==1.0.1